+++ lib/schema_dot_org/search_action.rb
@@ -5,7 +5,7 @@
 # Model the Schema.org `Thing > SearchAction`.  See http://schema.org/SearchAction
 #
 module SchemaDotOrg
-  class SearchAction < SchemaType
+  class SearchAction < SchemaDotOrg::SchemaType
     validated_attr :query_input, type: String, presence: true
     validated_attr :target,      type: String, presence: true
   end
